requirements.txt
altair==5.1.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
fillpdf==0.7.2
gitdb==4.0.10
GitPython==3.1.34
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathlib==1.0.1
pdf2image==1.16.3
pdfrw2==0.5.0
Pillow==9.5.0
protobuf==4.24.2
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
pymongo==4.5.0
Pympler==1.0.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
streamlit-option-menu==0.3.6
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validate-email-address==1
validators==0.22.0
watchdog==3.0.0
zipp==3.16.2
